graph-olap/packages/control-plane/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "control-plane"
version = "0.1.0"
description = "Graph OLAP Platform Control Plane"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "Proprietary"}
authors = [
    {name = "Graph OLAP Team"}
]
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pygtrie>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "alembic>=1.17.2",
    "structlog>=25.5.0",
    "httpx>=0.28.1",
    "google-cloud-storage>=3.7.0",
    "kubernetes>=34.1.0",
    "apscheduler>=3.11.2",
    "prometheus-client>=0.23.1",
    "tenacity>=9.1.2",
    "deepdiff>=8.6.1",
    "graph-olap-schemas>=1.0.0",
    "lib-auth>=1.0.0",
    "PyJWT[crypto]>=2.8.0",  # JWT validation with RS256 support
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
    "httpx>=0.28.1",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "testcontainers[postgres]>=4.0.0",
]

[project.scripts]
control-plane = "control_plane.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/control_plane"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: Integration tests (require database)",
    "e2e: End-to-end tests (require k3d cluster)",
]

[tool.coverage.run]
source = ["src/control_plane"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes (undefined names, unused imports)
    "I",      # isort (import ordering)
    "N",      # pep8-naming (naming conventions)
    "UP",     # pyupgrade (modern Python syntax)
    "B",      # flake8-bugbear (common bugs)
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify (code simplification)
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends)
    "SIM105", # use contextlib.suppress (sometimes less readable)
    "SIM117", # nested with statements (sometimes more readable)
    "N806",   # variable in function should be lowercase (NetworkX convention)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841", "ARG001", "ARG002"]  # unused variables/fixtures OK in tests

[tool.ruff.lint.isort]
known-first-party = ["control_plane"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_ignores = true
